requirements.txt
Django==3.1.2
djangorestframework==3.12.1
ipdb==0.13.4
psycopg2==2.8.6
